+++ lib/slices/pact/ui/generic/pact_break_creation_screen.dart
@@ -35,6 +35,7 @@ class _PactBreakCreationScreenState extends ConsumerState<PactBreakCreationScree
         unawaited(
           ref.read(analyticsServiceProvider).logScreenView(const PactBreakCreationAnalyticsScreen()),
         );
+        unawaited(ref.read(pactBreakCreationViewModelProvider(widget.pactId).notifier).load());
       }),
     );
   }